chore: 데이터플랫폼API Mock 파일 rename
